+++ test/_util.py
@@ -1,0 +1,14 @@
+"""
+Helper functions shared by tests.
+"""
+
+import asyncio
+
+
+def async_test(test):
+    """
+    Decorator to run async test methods.
+    """
+    def wrapper(*args, **kwargs):
+        asyncio.run(test(*args, **kwargs))
+    return wrapper